# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Celsius API",
    "version": "1.0",
    "description": "Celsius Mobile API Documentation"
  },
  "servers": [
    {
      "url": "https://celsiuspop.com",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/v1/network/claim": {
      "get": {
        "summary": "Fetch user's existing reward claims",
        "tags": [
          "MLM Network"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of reward claims"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Claim a level reward",
        "tags": [
          "MLM Network"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "level"
                ],
                "properties": {
                  "level": {
                    "type": "integer",
                    "enum": [
                      1,
                      3,
                      5,
                      7
                    ],
                    "description": "The Heaven level to claim reward for (1, 3, 5, 7)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reward claimed successfully"
          },
          "400": {
            "description": "Invalid level"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Level not completed yet"
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "Reward already claimed"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/v1/notifications/register": {
      "post": {
        "summary": "Register device for notifications",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "token"
                ],
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 10
                  },
                  "platform": {
                    "type": "string",
                    "enum": [
                      "ANDROID",
                      "IOS",
                      "WEB"
                    ],
                    "default": "ANDROID"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Device registered for notifications"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/orders/track": {
      "post": {
        "summary": "Track Order",
        "description": "Track an order using Order ID. Phone or Email is optional if the user is logged in.",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "BearerAuth": []
          },
          {}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "orderId"
                ],
                "properties": {
                  "orderId": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order details found"
          },
          "400": {
            "description": "Verification details required (if not logged in) or invalid request"
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    },
...
  },
...
}